KOGITO-1436: Refactoring + Tests
