+++ drools-wb-screens/drools-wb-scenario-simulation-editor/drools-wb-scenario-simulation-editor-kogito-runtime/src/test/java/org/drools/workbench/screens/scenariosimulation/webapp/client/editor/ScenarioSimulationEditorKogitoRuntimeScreenTest.java
@@ -15,27 +15,18 @@
  */
 package org.drools.workbench.screens.scenariosimulation.webapp.client.editor;
 
-import java.util.function.Consumer;
-
 import com.google.gwtmockito.GwtMockitoTestRunner;
 import org.drools.workbench.screens.scenariosimulation.kogito.client.editor.ScenarioSimulationEditorKogitoWrapper;
-import org.drools.workbench.screens.scenariosimulation.kogito.client.popup.ScenarioSimulationCreationPopupPresenter;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.ArgumentCaptor;
-import org.mockito.Captor;
 import org.mockito.Mock;
-import org.uberfire.backend.vfs.Path;
 import org.uberfire.client.mvp.PlaceManager;
-import org.uberfire.mvp.PlaceRequest;
-import org.uberfire.workbench.model.menu.Menus;
 
 import static org.drools.workbench.screens.scenariosimulation.webapp.client.editor.ScenarioSimulationEditorKogitoRuntimeScreen.SCENARIO_SIMULATION_KOGITO_RUNTIME_SCREEN_DEFAULT_REQUEST;
 import static org.junit.Assert.assertEquals;
 
 import static org.mockito.Mockito.spy;
-
 
 @RunWith(GwtMockitoTestRunner.class)
 public class ScenarioSimulationEditorKogitoRuntimeScreenTest {
@@ -43,15 +34,7 @@ public class ScenarioSimulationEditorKogitoRuntimeScreenTest {
     @Mock
     private PlaceManager placeManagerMock;
     @Mock
-    private PlaceRequest placeRequestMock;
-    @Mock
-    private Consumer<Menus> menusConsumerMock;
-    @Mock
     private ScenarioSimulationEditorKogitoWrapper scenarioSimulationEditorKogitoWrapperMock;
-    @Mock
-    private ScenarioSimulationCreationPopupPresenter scenarioSimulationCreationPopupPresenterMock;
-    @Captor
-    private ArgumentCaptor<Path> pathArgumentCaptor;
 
     private ScenarioSimulationEditorKogitoRuntimeScreen scenarioSimulationEditorKogitoRuntimeScreenSpy;
 
@@ -68,49 +51,9 @@ public class ScenarioSimulationEditorKogitoRuntimeScreenTest {
     public void getPlaceRequest() {
         assertEquals(SCENARIO_SIMULATION_KOGITO_RUNTIME_SCREEN_DEFAULT_REQUEST, scenarioSimulationEditorKogitoRuntimeScreenSpy.getPlaceRequest());
     }
+
 /*
-    @Test
-    public void onStartup() {
-        scenarioSimulationEditorKogitoRuntimeScreenSpy.onStartup(placeRequestMock);
-        verify(scenarioSimulationEditorKogitoWrapperMock, times(1)).onStartup(eq(placeRequestMock));
-    }
-
-    @Test
-    public void mayClose() {
-        scenarioSimulationEditorKogitoRuntimeScreenSpy.mayClose();
-        verify(scenarioSimulationEditorKogitoWrapperMock, times(1)).mayClose();
-    }
-
-    @Test
-    public void getTitleText() {
-        assertEquals(TITLE, scenarioSimulationEditorKogitoRuntimeScreenSpy.getTitleText() );
-    }
-
-    @Test
-    public void getTitle() {
-        scenarioSimulationEditorKogitoRuntimeScreenSpy.getTitle();
-        verify(scenarioSimulationEditorKogitoWrapperMock, times(1)).getTitle();
-    }
-
-    @Test
-    public void getWidget() {
-        scenarioSimulationEditorKogitoRuntimeScreenSpy.getWidget();
-        verify(scenarioSimulationEditorKogitoWrapperMock, times(1)).getWidget();
-    }
-
-    @Test
-    public void setMenus() {
-        scenarioSimulationEditorKogitoRuntimeScreenSpy.setMenus(menusConsumerMock);
-        verify(scenarioSimulationEditorKogitoWrapperMock, times(1)).setMenus(eq(menusConsumerMock));
-    }
-
-    @Test
-    public void getContent() {
-        scenarioSimulationEditorKogitoRuntimeScreenSpy.getContent();
-        verify(scenarioSimulationEditorKogitoWrapperMock, times(1)).getContent();
-    }
-
-    @Test
+   @Test
     public void setContent() {
         scenarioSimulationEditorKogitoRuntimeScreenSpy.setContent("path/file.scesim", "value");
         verify(scenarioSimulationEditorKogitoRuntimeScreenSpy, never()).newFile(any());
@@ -165,9 +108,5 @@ public class ScenarioSimulationEditorKogitoRuntimeScreenTest {
         assertEquals("path\\", pathArgumentCaptor.getValue().toURI());
     }
 
-    @Test
-    public void isDirty() {
-        scenarioSimulationEditorKogitoRuntimeScreenSpy.isDirty();
-        verify(scenarioSimulationEditorKogitoWrapperMock, times(1)).isDirty();
-    }*/
+*/
 }
