+++ drools-wb-screens/drools-wb-scenario-simulation-editor/drools-wb-scenario-simulation-editor-kogito-webapp-common/src/test/java/org/drools/workbench/screens/scenariosimulation/webapp/client/editor/AbstractScenarioSimulationEditorKogitoScreenTest.java
@@ -15,21 +15,23 @@
  */
 package org.drools.workbench.screens.scenariosimulation.webapp.client.editor;
 
+import java.util.function.Consumer;
+
 import com.google.gwtmockito.GwtMockitoTestRunner;
 import org.drools.workbench.screens.scenariosimulation.kogito.client.editor.ScenarioSimulationEditorKogitoWrapper;
-import org.drools.workbench.screens.scenariosimulation.kogito.client.popup.ScenarioSimulationCreationPopupPresenter;
-import org.jboss.errai.common.client.api.RemoteCallback;
 import org.junit.Before;
+import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.ArgumentCaptor;
-import org.mockito.Captor;
 import org.mockito.Mock;
-import org.uberfire.backend.vfs.Path;
-import org.uberfire.mvp.Command;
 import org.uberfire.mvp.PlaceRequest;
+import org.uberfire.workbench.model.menu.Menus;
 
+import static org.drools.workbench.screens.scenariosimulation.webapp.client.editor.AbstractScenarioSimulationEditorKogitoScreen.TITLE;
+import static org.junit.Assert.assertEquals;
+import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.spy;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 
 @RunWith(GwtMockitoTestRunner.class)
 public class AbstractScenarioSimulationEditorKogitoScreenTest {
@@ -37,13 +39,9 @@ public class AbstractScenarioSimulationEditorKogitoScreenTest {
     @Mock
     private ScenarioSimulationEditorKogitoWrapper scenarioSimulationEditorKogitoWrapperMock;
     @Mock
-    private ScenarioSimulationCreationPopupPresenter scenarioSimulationCreationPopupPresenterMock;
-    @Captor
-    private ArgumentCaptor<Command> createCommandCaptor;
-    @Captor
-    private ArgumentCaptor<RemoteCallback> remoteCallbackArgumentCaptor;
-    @Captor
-    private ArgumentCaptor<Path> pathArgumentCaptor;
+    private PlaceRequest placeRequestMock;
+    @Mock
+    private Consumer<Menus> menusConsumerMock;
 
     private AbstractScenarioSimulationEditorKogitoScreen abstractScenarioSimulationEditorKogitoScreenSpy;
 
@@ -60,76 +58,59 @@ public class AbstractScenarioSimulationEditorKogitoScreenTest {
             }
 
         });
-        when(scenarioSimulationCreationPopupPresenterMock.getSelectedPath()).thenReturn("selected");
-    }
-
-    /*
-    @Test
-    public void newFileEmptySelectedType() {
-        Path path = new PathFactory.PathImpl("file.scesim", "path/");
-        abstractScenarioSimulationEditorKogitoScreenSpy.newFile(path);
-        verify(scenarioKogitoCreationPopupPresenterMock, times(1)).show(isA(String.class), createCommandCaptor.capture());
-        createCommandCaptor.getValue().execute();
-        verify(scenarioKogitoCreationPopupPresenterMock, times(1)).getSelectedType();
-        verify(abstractScenarioSimulationEditorKogitoScreenSpy, times(1)).showPopover(eq("ERROR"), eq("Missing selected type"));
-        verify(scenarioSimulationBuilderMock, never()).populateScenarioSimulationModel(isA(ScenarioSimulationModel.class),
-                                                                                       eq(ScenarioSimulationModel.Type.RULE),
-                                                                                       eq(""),
-                                                                                       remoteCallbackArgumentCaptor.capture());
     }
 
     @Test
-    public void newFileEmptySelectedDMNPath() {
-        when(scenarioKogitoCreationPopupPresenterMock.getSelectedPath()).thenReturn(null);
-        when(scenarioKogitoCreationPopupPresenterMock.getSelectedType()).thenReturn(ScenarioSimulationModel.Type.DMN);
-        Path path = new PathFactory.PathImpl("file.scesim", "path/");
-        abstractScenarioSimulationEditorKogitoScreenSpy.newFile(path);
-        verify(scenarioKogitoCreationPopupPresenterMock, times(1)).show(isA(String.class), createCommandCaptor.capture());
-        createCommandCaptor.getValue().execute();
-        verify(scenarioKogitoCreationPopupPresenterMock, times(1)).getSelectedType();
-        verify(abstractScenarioSimulationEditorKogitoScreenSpy, times(1)).showPopover(eq("ERROR"), eq("Missing dmn path"));
-        verify(scenarioSimulationBuilderMock, never()).populateScenarioSimulationModel(isA(ScenarioSimulationModel.class),
-                                                                                       eq(ScenarioSimulationModel.Type.RULE),
-                                                                                       eq(""),
-                                                                                       remoteCallbackArgumentCaptor.capture());
+    public void onStartup() {
+        abstractScenarioSimulationEditorKogitoScreenSpy.onStartup(placeRequestMock);
+        verify(scenarioSimulationEditorKogitoWrapperMock, times(1)).onStartup(eq(placeRequestMock));
     }
 
     @Test
-    public void newFileRule() {
-        when(scenarioKogitoCreationPopupPresenterMock.getSelectedType()).thenReturn(ScenarioSimulationModel.Type.RULE);
-        Path path = new PathFactory.PathImpl("file.scesim", "path/");
-        abstractScenarioSimulationEditorKogitoScreenSpy.newFile(path);
-        verify(scenarioKogitoCreationPopupPresenterMock, times(1)).show(isA(String.class), createCommandCaptor.capture());
-        createCommandCaptor.getValue().execute();
-        verify(scenarioKogitoCreationPopupPresenterMock, times(1)).getSelectedType();
-        verify(scenarioSimulationBuilderMock, times(1)).populateScenarioSimulationModel(isA(ScenarioSimulationModel.class),
-                                                                                                             eq(ScenarioSimulationModel.Type.RULE),
-                                                                                                             eq(""),
-                                                                                                             remoteCallbackArgumentCaptor.capture());
-        remoteCallbackArgumentCaptor.getValue().callback("");
-        verify(abstractScenarioSimulationEditorKogitoScreenSpy, times(1)).saveFile(pathArgumentCaptor.capture(), isA(String.class));
-        verify(scenarioSimulationEditorKogitoWrapperMock, times(1)).gotoPath(eq(pathArgumentCaptor.getValue()));
-        verify(scenarioSimulationEditorKogitoWrapperMock, times(1)).setContent(eq("path/file.scesim"), isA(String.class));
-        assertEquals("path/", pathArgumentCaptor.getValue().toURI());
-        assertEquals("file.scesim", pathArgumentCaptor.getValue().getFileName());
+    public void mayClose() {
+        abstractScenarioSimulationEditorKogitoScreenSpy.mayClose();
+        verify(scenarioSimulationEditorKogitoWrapperMock, times(1)).mayClose();
     }
 
     @Test
-    public void newFileDMN() {
-        when(scenarioKogitoCreationPopupPresenterMock.getSelectedType()).thenReturn(ScenarioSimulationModel.Type.DMN);
-        Path path = new PathFactory.PathImpl("file.scesim", "path/");
-        abstractScenarioSimulationEditorKogitoScreenSpy.newFile(path);        verify(scenarioKogitoCreationPopupPresenterMock, times(1)).show(isA(String.class), createCommandCaptor.capture());
-        createCommandCaptor.getValue().execute();
-        verify(scenarioKogitoCreationPopupPresenterMock, times(1)).getSelectedType();
-        verify(scenarioSimulationBuilderMock, times(1)).populateScenarioSimulationModel(isA(ScenarioSimulationModel.class),
-                                                                                        eq(ScenarioSimulationModel.Type.DMN),
-                                                                                        eq("selected"),
-                                                                                        remoteCallbackArgumentCaptor.capture());
-        remoteCallbackArgumentCaptor.getValue().callback("");
-        verify(abstractScenarioSimulationEditorKogitoScreenSpy, times(1)).saveFile(pathArgumentCaptor.capture(), isA(String.class));
-        verify(scenarioSimulationEditorKogitoWrapperMock, times(1)).gotoPath(eq(pathArgumentCaptor.getValue()));
-        verify(scenarioSimulationEditorKogitoWrapperMock, times(1)).setContent(eq("path/file.scesim"), isA(String.class));
-        assertEquals("path/", pathArgumentCaptor.getValue().toURI());
-        assertEquals("file.scesim", pathArgumentCaptor.getValue().getFileName());
-    }*/
+    public void getTitleText() {
+        assertEquals(TITLE, abstractScenarioSimulationEditorKogitoScreenSpy.getTitleText() );
+    }
+
+    @Test
+    public void getTitle() {
+        abstractScenarioSimulationEditorKogitoScreenSpy.getTitle();
+        verify(scenarioSimulationEditorKogitoWrapperMock, times(1)).getTitle();
+    }
+
+    @Test
+    public void getWidget() {
+        abstractScenarioSimulationEditorKogitoScreenSpy.getWidget();
+        verify(scenarioSimulationEditorKogitoWrapperMock, times(1)).getWidget();
+    }
+
+    @Test
+    public void setMenus() {
+        abstractScenarioSimulationEditorKogitoScreenSpy.setMenus(menusConsumerMock);
+        verify(scenarioSimulationEditorKogitoWrapperMock, times(1)).setMenus(eq(menusConsumerMock));
+    }
+
+    @Test
+    public void getContent() {
+        abstractScenarioSimulationEditorKogitoScreenSpy.getContent();
+        verify(scenarioSimulationEditorKogitoWrapperMock, times(1)).getContent();
+    }
+
+    @Test
+    public void setContent() {
+        abstractScenarioSimulationEditorKogitoScreenSpy.setContent("fullPath", "content");
+        verify(scenarioSimulationEditorKogitoWrapperMock, times(1)).setContent(eq("fullPath"),
+                                                                                                    eq("content"));
+    }
+
+    @Test
+    public void isDirty() {
+        abstractScenarioSimulationEditorKogitoScreenSpy.isDirty();
+        verify(scenarioSimulationEditorKogitoWrapperMock, times(1)).isDirty();
+    }
 }
